2.2.6: fixed dbstats

+++ src/main/java/de/hablijack/greenhouse/api/pojo/TableSize.java
@@ -1,14 +1,12 @@
 package de.hablijack.greenhouse.api.pojo;
 
-import java.math.BigInteger;
-
 public class TableSize {
 
-  private BigInteger measurementSizeByte;
+  private Long measurementSizeByte;
 
-  private BigInteger relayLogSizeByte;
+  private Long relayLogSizeByte;
 
-  public TableSize(BigInteger measurementSizeByte, BigInteger relayLogSizeByte) {
+  public TableSize(Long measurementSizeByte, Long relayLogSizeByte) {
     this.measurementSizeByte = measurementSizeByte;
     this.relayLogSizeByte = relayLogSizeByte;
   }
@@ -16,19 +14,19 @@ public class TableSize {
   public TableSize() {
   }
 
-  public BigInteger getMeasurementSizeByte() {
+  public Long getMeasurementSizeByte() {
     return measurementSizeByte;
   }
 
-  public void setMeasurementSizeByte(BigInteger measurementSizeByte) {
+  public void setMeasurementSizeByte(Long measurementSizeByte) {
     this.measurementSizeByte = measurementSizeByte;
   }
 
-  public BigInteger getRelayLogSizeByte() {
+  public Long getRelayLogSizeByte() {
     return relayLogSizeByte;
   }
 
-  public void setRelayLogSizeByte(BigInteger relayLogSizeByte) {
+  public void setRelayLogSizeByte(Long relayLogSizeByte) {
     this.relayLogSizeByte = relayLogSizeByte;
   }
 }
